Try improving performance by splitting BreathGuide in multiple different widgets, avoiding rebuilds.
