+++ lib/ui/painters/breath_pattern_painter.dart
@@ -17,7 +17,7 @@ class BreathPatternPainter extends CustomPainter {
 
   @override
   void paint(Canvas canvas, Size size) {
-    print('Repainted pattern!');
+    print('Repainted $BreathPatternPainter!');
 
     // Calculate the center horizontal.
     final centerX = size.width / 2;
